Replace LogtailAppender by Logtail4j implementation
