+++ src/test/java/com/logtail/logback/LogtailAppenderDecorator.java
@@ -1,42 +1,23 @@
-
 package com.logtail.logback;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import jakarta.ws.rs.core.Response;
+import java.io.IOException;
 
 public class LogtailAppenderDecorator extends LogtailAppender {
-
-    private Throwable exception;
-
-    private Response response;
-
-    private LogtailResponse logtailResponse;
-
+    private Exception exception;
+    private LogtailResponse response;
     protected int apiCalls = 0;
 
     @Override
-    protected Response callIngestApi(String jsonData) {
-
+    protected LogtailResponse callHttpURLConnection() throws IOException {
         try {
-
-            this.response = super.callIngestApi(jsonData);
+            this.response = super.callHttpURLConnection();
             apiCalls++;
-            return response;
 
-        } catch (Throwable t) {
-
-            this.exception = t;
-            throw t;
-
+            return this.response;
+        } catch (Exception e) {
+            this.exception = e;
+            throw e;
         }
-
-    }
-
-    @Override
-    protected LogtailResponse convertResponseToObject(Response response) throws JsonProcessingException, JsonMappingException {
-        this.logtailResponse = super.convertResponseToObject(response);
-        return this.logtailResponse;
     }
 
     public boolean hasException() {
@@ -44,23 +25,18 @@ public class LogtailAppenderDecorator extends LogtailAppender {
     }
 
     public boolean hasError() {
-        return this.logtailResponse != null && this.logtailResponse.getError() != null;
+        return this.response != null && this.response.getError() != null;
     }
 
     public boolean isOK() {
         return this.response != null && this.response.getStatus() == 202;
     }
 
-    public Throwable getException() {
+    public Exception getException() {
         return exception;
     }
 
-    public Response getResponse() {
+    public LogtailResponse getResponse() {
         return response;
     }
-
-    public LogtailResponse getLogtailResponse() {
-        return logtailResponse;
-    }
-
 }
